added tabs model & schema
