+++ bucket-pasta-server/src/models/tabs/tabs-model.js
@@ -1,14 +1,14 @@
 'use strict';
 
-const mongoose = require('mongoose');
-let pasties = require('../pasties/pasties-schema.js');
+const Model = require('../mongo.js.js');
+const schema = require('./tabs-schema.js');
 
-const tab = mongoose.Schema({
-    displayName: String,
-    theme: String,
-    type: String,
-    pasties: [pasties] //not sure if this is doing what I think. It should hopefully read this as an array of objects that look like the pasties schema.
-})
+/**
+ * Class representing a Category.
+ * @extends Model
+ */
+class Tabs extends Model {
+  constructor() { super(schema); }
+}
 
-
-module.exports = mongoose.model('Tab', tab);
+module.exports = Tabs;
